Remove reliance of issues package for general types.
Remove profile link from sgUser since it's removed upstream.

Remove CopierFrom optional interface for now, it's not implemented and not
used. It was out of date. It can be re-introduced on demand when needed.

Add TODO notes for general structs that can be factored out of notification
center app when it's appropriate. It's not worth to do now.