Add MarkAllRead method.
Change githubapi RepoSpec expectation to be of form
"github.com/owner/repo", rather than just "owner/repo". It was already
partially that way, and it seems like a good direction, but we'll see.

Preserve AppID and ThreadID in fs implementation because it becomes
needed for MarkAllRead where the AppID and ThreadID are not known.

Remove InternalService and keep only Service, because InternalService
was unhelpful (it wasn't sufficient, you'd also need MarkAllRead which
then needs to appear in both).