githubapi: Perform pagination in Count implementation.
This is more accurate. It correctly reports the total number of
notifications, even if there are more of them that fit on one page (50
is the default page size for this endpoint, it seems).

It is also more efficient, because we're receiving at most one
notification in the response, rather than the default page size 50).