githubapi: Use GitHub's global notification IDs.
This is possible because it appears we can parse the notification ID
(a numeric string like "230400425") from GitHub into a uint64 reliably.

It greatly simplifies the MarkRead implementation.

Fixes strange rare issues with marking some threads as read, seemingly
due to incorrect caching. It may be worthwhile to try to get to the
bottom of the underlying issue regardless (if it's a bug in the
github.com/gregjones/httpcache library caching behavior).

Also improves performance of MarkRead method, due to making fewer API
calls.