Revert "githubapi: Use GitHub's global notification IDs."
This reverts commit ed5c28f99c6a04dc82dd3887e2cac13562fa1602.

Unfortunately, this optimization is not possible, at least not as is.
We need to return actual issue IDs as ThreadIDs in List, so that
issuesapp.augmentUnread works correctly. It expects the issue IDs
to match thread IDs for "Issue" thread type.

Another attempt at this optimization can be made in the future, without
breaking external (e.g., issuesapp) behavior. It'll be a good idea to
first add a test for that behavior, so it doesn't regress.