githubapi: Improve MarkRead handling of global notification IDs.
The previous handling of "Commit", "Release", "RepositoryInvitation"
thread types was not correct. Fix that, and optimize it by handling it
directly as first step.

Make global notification ID parse errors fatal, because it seems they
never happen, so I want to make it very visible if they ever do.

Update comment about why we can't use global notification IDs for
issues, pull requests thread types.

Report error in MarkRead if the specified thread ID was not marked read
(because it couldn't be found).