githubapi: Try again without cache in MarkRead if notification not found.
There are sometimes caching issues causing stale repository notifications
to be retrieved from cache, and a legitimate existing notification is not
marked read. So fall back to skipping cache, if we can't find a notification
and the response we got was from cache (rather than origin server).

Log details when this occurs to help investigate the issue and determine
its cause.