githubapi: Use GitHub GraphQL API v4, make better queries.
Use GraphQL API to get issue/PR state, and the author/comment ID of the
latest comment. If there are no comments, then use the author of the
issue/PR itself.

It's possible to do this more efficiently (fewer requests) with GraphQL
API rather than REST API. For instance, it took two separate requests
with REST to get the issue state and the author; now it takes just one.

As a result, we also rely less on the notification's LatestCommentURL,
which is a good thing, because that field seems to be unreliable. It's
not always present or has the expected value; instead, it often equals
to regular issue/PR URL and thus doesn't specify the latest comment.

Using GraphQL API here also makes it possible to fetch more information
efficiently in the future (e.g., the latest comment body, etc.).

Fix HTMLURL links to individual comments. They have fragment
"#issuecomment-{ID}" rather than "#comment-{ID}".

Document expensive API calls, and opportunities for future improvements.