Rename, reorder AppID to ThreadType; remove RepoURL.
ThreadType is a much more clear and logical name, instead of AppID.
It better communicates its purpose of giving a namespace to ThreadIDs,
which could otherwise conflict.

Placing it second, after RepoSpec, and before ThreadID, is the most
logical and consistent position. That way, it's a proper hierarchy,
from highest level to lowest level.

RepoURL is redundant. It can be replaced with "https://"+RepoSpec.URI.

Resolves the following longstanding TODO:

	// TODO: Rename AppID to "ThreadType" and shuffle its order to make
	//       it more clear: (RepoSpec, ThreadType, ThreadID).