githubapi: Perform pagination in MarkRead.
If there are more than 50 notifications (the default page size
for notifications) in a repository, target notification might
not get found, because it would be on page 2 or later. So it's
important to paginate over all pages, and look for the target
notification everywhere.

Set page size to maximum of 100, to reduce number of API calls.

Refactor such that as soon as the target notification is found,
it gets marked read and MarkRead returns. This is simpler.